lesson 5 Filter fix bug massive/object in GroupList + add component itemGroupList
