+++ src/app/components/itemGroupList.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+
+const listItem = (itemElement, valueProperty, contentProperty, onItemSelect, selectedItem) => {
+    return (
+        <li
+            key={itemElement[valueProperty]}
+            className={
+                'list-group-item' +
+                (itemElement === selectedItem ? ' active' : '')
+            }
+            onClick={() => onItemSelect(itemElement)}
+            role={'button'}
+        >
+            {itemElement[contentProperty]}
+        </li>
+    );
+};
+
+export default listItem;
